+++ src/main/java/com/grup14/luterano/entities/HistorialCurso.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -28,7 +29,7 @@ public class HistorialCurso {
     private LocalDate fechaDesde;
     private LocalDate fechaHasta;
 
-    private float promedio; // promedio global del curso
+    private BigDecimal promedio; // promedio global del curso
 
     @OneToMany(mappedBy = "historialCurso", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<HistorialMateria> historialMaterias = new ArrayList<>();
